#18 [Refactor] store id를 int에서 long으로 변경

+++ src/main/java/com/mpnp/baechelin/review/dto/ReviewResponseDto.java
@@ -16,7 +16,7 @@ import java.util.stream.Collectors;
 @Getter @Setter
 public class ReviewResponseDto {
     private int reviewId; // storeId
-    private int storeId;
+    private long storeId;
     private int userId;
     private Double point;
